Added boolean parameter to mask label img if input data has nodata value (#39)
* Added boolean parameter to mask label img if input data has nodata value

+++ images_to_samples.py
@@ -170,7 +170,7 @@ def vector_to_raster(vector_file, attribute_name, new_raster):
 
 
 def main(bucket_name, data_path, samples_size, num_classes, number_of_bands, csv_file, samples_dist,
-         remove_background, mask_input_image):
+         remove_background, mask_input_image, mask_reference):
     gpkg_file = []
     if bucket_name:
         s3 = boto3.resource('s3')
@@ -222,22 +222,29 @@ def main(bucket_name, data_path, samples_size, num_classes, number_of_bands, csv
             info['gpkg'] = info['gpkg'].split('/')[-1]
         assert_band_number(info['tif'], number_of_bands)
 
-        tmp_label_raster = create_new_raster_from_base(info['tif'], tmp_label_name, 1)
         value_field = info['attribute_name']
         validate_num_classes(info['gpkg'], num_classes, value_field)
-        vector_to_raster(info['gpkg'], info['attribute_name'], tmp_label_raster)
-
-        tmp_label_raster = None
 
         # Mask zeros from input image into label raster.
-        masked_array = mask_image(image_reader_as_array(info['tif']), image_reader_as_array(tmp_label_name))
-        create_new_raster_from_base(info['tif'], label_name, 1, masked_array)
+        if mask_reference:
+            tmp_label_raster = create_new_raster_from_base(info['tif'], tmp_label_name, 1)
+            vector_to_raster(info['gpkg'], info['attribute_name'], tmp_label_raster)
+            tmp_label_raster = None
+
+            masked_array = mask_image(image_reader_as_array(info['tif']), image_reader_as_array(tmp_label_name))
+            create_new_raster_from_base(info['tif'], label_name, 1, masked_array)
+
+            os.remove(tmp_label_name)
+
+        else:
+            label_raster = create_new_raster_from_base(info['tif'], label_name, 1)
+            vector_to_raster(info['gpkg'], info['attribute_name'], label_raster)
+            label_raster = None
+
         # Mask zeros from label raster into input image.
         if mask_input_image:
             masked_img = mask_image(image_reader_as_array(label_name), image_reader_as_array(info['tif']))
             create_new_raster_from_base(label_name, info['tif'], number_of_bands, masked_img)
-
-        os.remove(tmp_label_name)
 
         if info['dataset'] == 'trn':
             out_file = tmp_trn_hdf5
@@ -309,4 +316,5 @@ if __name__ == '__main__':
          params['sample']['prep_csv_file'],
          params['sample']['samples_dist'],
          params['sample']['remove_background'],
-         params['sample']['mask_input_image'])
+         params['sample']['mask_input_image'],
+         params['sample']['mask_reference'])
